# src/api/v1/schemas/room.go
package schemas

type RoomCreateSchema struct {
	Name          string `json:"name"`
	IdComputerLab string `json:"id_computer_lab"`
}

type RoomUpdateSchema struct {
	// Metadata

	// Data
	Name  string `json:"name"`
	Index int    `json:"index"`
}